# Optimize map Presto function for constant keys