Summary:
A common pattern in queries to apply 'map' function to a constant array of keys:

```
SELECT map(array[1, 2, 3], array[a, b, c])) ...
```

In this case, a lot of CPU cycles are spent checking for duplicate map keys again and again for each row. It is sufficient to check once.

Benchmark results before

```
============================================================================
[...]prestosql/benchmarks/MapBenchmark.cpp     relative  time/iter   iters/s
============================================================================
constantKeys                                              256.17ns     3.90M
flatKeys                                                  237.76ns     4.21M
```

and after the optimization

```
============================================================================
[...]prestosql/benchmarks/MapBenchmark.cpp     relative  time/iter   iters/s
============================================================================
constantKeys                                               71.09ns    14.07M
flatKeys                                                  248.60ns     4.02M
```

Differential Revision: D45720372